+++ spotbugs/src/test/java/edu/umd/cs/findbugs/util/MutableClassesTest.java
@@ -1,5 +1,8 @@
 package edu.umd.cs.findbugs.util;
 
+import java.io.Externalizable;
+import java.io.ObjectInput;
+import java.io.ObjectOutput;
 import java.io.ObjectStreamException;
 import java.io.Serializable;
 import javax.annotation.concurrent.Immutable;
@@ -69,7 +72,7 @@ public class MutableClassesTest {
     }
 
     public static class Immutable {
-        private int n;
+        private final int n;
         private static Immutable immutable;
 
         public Immutable(int n) {
@@ -120,5 +123,53 @@ public class MutableClassesTest {
     public void TestImmutableWriteReplace() {
         Assert.assertFalse(MutableClasses.mutableSignature(
                 "Ledu/umd/cs/findbugs/util/MutableClassesTest$ImmutableWriteReplace;"));
+    }
+
+    public static final class MutableWriteExternal {
+        void writeExternal(ObjectOutput out) {
+            // Does not matter
+        }
+    }
+
+    public static class MutableWriteExternalSig implements Externalizable {
+        @Override
+        public void writeExternal(ObjectOutput out) {
+            // Does not matter
+        }
+
+        @Override
+        public void readExternal(ObjectInput in) {
+            // Does not matter
+        }
+
+        void writeExternal() {
+            // Does not match signature
+        }
+    }
+
+    @Test
+    public void TestMutableWriteExternal() {
+        Assert.assertTrue(MutableClasses.mutableSignature(
+                "Ledu/umd/cs/findbugs/util/MutableClassesTest$MutableWriteExternal;"));
+        Assert.assertTrue(MutableClasses.mutableSignature(
+                "Ledu/umd/cs/findbugs/util/MutableClassesTest$MutableWriteExternalSig;"));
+    }
+
+    public static final class ImmutableWriteExternal implements Externalizable {
+        @Override
+        public void writeExternal(ObjectOutput out) {
+            // Does not matter
+        }
+
+        @Override
+        public void readExternal(ObjectInput in) {
+            // Does not matter
+        }
+    }
+
+    @Test
+    public void TestImmutableWriteExternal() {
+        Assert.assertFalse(MutableClasses.mutableSignature(
+                "Ledu/umd/cs/findbugs/util/MutableClassesTest$ImmutableWriteExternal;"));
     }
 }