+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Currently the versioning policy of this project follows [Semantic Versioning v2.
 `SW_SWING_METHODS_INVOKED_IN_SWING_THREAD` ([#1664](https://github.com/spotbugs/spotbugs/pull/1664))
 - Wrong description of the `SE_TRANSIENT_FIELD_OF_NONSERIALIZABLE_CLASS` ([#1664](https://github.com/spotbugs/spotbugs/pull/1664))
 - Do not consider java.io.Serializable with writeReplace() as mutable type ([#1704](https://github.com/spotbugs/spotbugs/pull/1704))
+- Do not Externalizable.writeExternal() as a mutating method ([]())
 
 ## 4.4.1 - 2021-09-07
 ### Changed